hw/ipmi: Add an IPMI external host device
The IPMI external host device works for Baseband Management Controller
(BMC) emulations. It works as a representation of a host class that
connects to a given BMC.  It can connect to a real host hardware or a
emulated or simulated host device. In particular it can connect to a
host QEMU instance with device ipmi-bmc-extern.

For more details of IPMI host device in BMC emulation, see
docs/specs/ipmi.rst.

GitWatcher: ignore
Google-Bug-Id: 153654878
Change-Id: If86e01ffeb90010041f7b322bd4613bfb8858a2b
Signed-off-by: Hao Wu <wuhaotsh@google.com>
Original-Change-Id: http://go/team-change-id/Ied8ccd3843a34081824bd44f327feffd25753268