hw/misc: Add Nuvoton's PCI Mailbox Module
The PCI Mailbox Module is a high-bandwidth communication module
between a Nuvoton BMC and CPU. It features 16KB RAM that are both
accessible by the BMC and core CPU. and supports interrupt for
both sides.

This patch implements the BMC side of the PCI mailbox module.
Communication with the core CPU is emulated via a chardev and
will be in a follow-up patch.

GitWatcher: ignore
Google-Bug-Id: 262936693
Signed-off-by: Hao Wu <wuhaotsh@google.com>
Original-Change-Id: http://go/team-change-id/If13212bfc9a5e75532e49b554f0155ed8210fc0d
Change-Id: If13212bfc9a5e75532e49b554f0155ed8210fc0d