hw/pci: create guest only pci device
Not likely that we'll upstream this, but this is similar to the
device we use for Vanadium that can become a pci configuration
space for any device we need within reason.  Obviously it doesn't
hold up to scrutiny, but often this is enough.

Patrick Venture (1):
      hw/pci: create guest only pci device

Nabih Estefan (1):
      hw/pci/guest-only-pci: Add non-functional BAR Space

GitWatcher: ignore
Google-Bug-Id: 195554472
Google-Bug-Id: 237108612
Google-Bug-Id: 378933803
FIXED=195554472
PiperOrigin-RevId: 457091214
Signed-off-by: Patrick Venture <venture@google.com>
Signed-off-by: Nabih Estefan <nabihestefan@google.com>
Change-Id: I672b050fb8f724a3b3cc5d9659a3b94f26b68bb3