hw/timer/sse-timer: Add CNTFRQ reset property
This can have a non-zero reset value, and cause the kernel to complain
about a CNTFRQ mismatch if TF-A (or firmware in general) does not
initialize it (because it expects the value to be non-zero out of
reset).

To fix this, we'll just add an object property that people can use to
initialize the CNTFRQ reset value.

GitWatcher: ignore
Google-Bug-Id: 277794748
Signed-off-by: Joe Komlodi <komlodi@google.com>
Change-Id: Ie97d081403e8272d9186fefcf60996fe3a3d8921