hw/i2c: smbus: mux: Reset SMBusDevice state on reset
When a reset happens, both the SMBusDevice and PCA954x class do their
variable resetting on an enter reset. Because of this, only the PCA954x
has its reset called, which can leave the SMBusDevice in a bad state if
it was in the middle of a transaction.

To fix this we add parent reset functions for the SMBusDevice class, and
have the mux class invoke it when it resets.

GitWatcher: ignore
Google-Bug-Id: 231153633
Signed-off-by: Joe Komlodi <komlodi@google.com>
Change-Id: I50bb22b36ba2547e56fd8311b6a5bf8644612710