hw/i3c/svc-i3c: Add I3C data reception
I3C data reception is treated mostly the same way as I2C data reception,
except I3C targets have the way to end transactions early.

CCCs are treated the same as other transactions, so we do not need to
keep track if the transaction is a CCC or not.

As part of this we also add I3C STARTs and STOPs.

GitWatcher: ignore
Google-Bug-Id: 308692346
Change-Id: I62f8613b23cbc54ec7928539fa7cc3ebd532a478
Signed-off-by: Joe Komlodi <komlodi@google.com>