hw/i3c/svc-i3c: Add ENTDAA support
ENTDAA behaves differently from other transfers. The user must program
the controller into ENTDAA mode, and periodically send addresses for the
controller to assign to targets.

Once the controller gets NACKed when issuing reSTARTs to see who else is
on the bus during ENTDAA, the user should stop the transaction, which
ends the ENTDAA process.

Google-Bug-Id: 308692346
Change-Id: I0bd3828e9cffb2b674d1f993abf27dc7a21e42e7
Signed-off-by: Joe Komlodi <komlodi@google.com>