hw/i3c/svc-i3c: Add IBI handling
For now we only support automatic IBI ACKing/NACKing, which is what the
driver also supports.

The controller supports a mode where control is given to softwre to
decide whether or not it should ACK or NACK the IBI, however
supporting this in QEMU is quite challening.
IBIs are done in one atomic transaction, so there is no chance for the
guest to be able to ACK or NACK the IBI. By the time control is given
back to the guest, the iBI would have finished.

Since we don't support manual IBI ACKing/NACKing, we decide to NACK
because ACKing an IBI that the driver doesn't expect may cause driver
crashes and other undefined driver behavior.

Google-Bug-Id: 308692346
Change-Id: I28f929948ae005bf78b0bb2f9c1b3fb2878eb539
Signed-off-by: Joe Komlodi <komlodi@google.com>