AEMU: Enable meson generator on native windows and other build mods
This enables the meson generator to run under windows. In order to
accomplish this we have:

- Removed all dependencies on Awk and Bash
- Fixed a series of small issues in the meson scripts
- Created python versions of existing shell scripts

All the source generation scripts now only require a python interpreter,
enabling you to configure the build under windows.

Note: you will still need a clang-cl compiler that is properly
configured to handle the actual compilation.

Note that this is mostly cherry-picked from main-emu-next-dev branch.

Bug: 417477722
Change-Id: Ic1bfce4b80c3a4fc2f622ef49e508189f60a9569