hw/i3c/remote-i3c: Fix receiving IBI bytes during address matching
It was possible for us to receive a REMOTE_I3C_IBI byte while expecting
a response to an address match request. Because we treated non-zero
bytes as an ACK, we would read this as an ACK and would cause the
following problems to happen:

1. QEMU would erroneously believe it had control of the bus, when it did
   not because the IBI came first
2. The next byte sent by QEMU would be interpreted as the IBI ACK/NACK
   byte.

To deal with this, we make the ACK/NACK byte more strict so it can only
be a 1 or 0, and if we receive a REMOTE_I3C_IBI byte, we pass it to the
IBI RX state machine.
We also add an additional check to see if an IBI is in progress, and
NACK right away if it is. This doesn't fix any bugs, but is a small
optimization and saves a socket read/write call.

It's worth noting that this problem occurs because we receive socket
data on 2 threads, the MMIO one and the chardev one.
I explored having only the chardev thread be the recipient, and use
semaphores to notify the MMIO thread when data reception finished,
however this had a couple drawbacks:

1. The code became quite complex. Each Receive function needed its own
   state machine that could handle byte-by-byte packets and we had to be
   very careful about maintaining internal state. This isn't bad on its
   own, however...
2. This made things 700-1000x slower, maybe more. This is because the
   delay betweeen qemu_sem_post() and the MMIO thread that called
   qemu_sem_wait() getting scheduled again went from around <1ms to
   700+ms.

Unfortunately the above delay made this solution infeasible, so we opt
for the stricter check solution.

Google-Bug-Id: 331418723
Change-Id: I5a2f5a060d298d0303dff3b21f915472d5a4f851
Signed-off-by: Joe Komlodi <komlodi@google.com>