args: Parse -L option during first pass
The BIOS path on X86 is needed before the second pass of parameters is
parsed. This makes it impossible to use a custom BIOS path.

This commit duplicates the parsing of the -L option to the first pass of
parameters, allowing the BIOS path to be set correctly.

Note that this is cherry-picked from main-emu-next-dev branch.

Bug: 417477722
Test: Can launch qemu with custom bios paths
Change-Id: I2ec6ddc2e122db4be428a87e32d454e970777761