Add mcux-soc-svd subproject
Add mcux-soc-svd subproject that contains SVD files that are going to
be used to generate C header with register layout definitions and
other helpers to create device models.

Guard the subproject by a configuration option since it is rarely
going to be used - whenever new headers will be generated. It is
unlikely that already generated headers will be updated, with the
exception of minor hardware revisions.

Also export the rt595 SVD file which is going to be used by subsequent
patches.

KOKORO_SKIP_CHECKPATCH
GitWatcher: ignore
Change-Id: I7713a37fcc51b414d755a85b1cfe580fd1c3641c
Signed-off-by: Octavian Purdila <tavip@google.com>