hw/misc: add support for flexcomm
Add support for NXP's FLEXCOMM device model.

FLEXCOMM is a generic serial communication module which support
multiple functions: UART, SPI and I2C. These are configurable at
runtime.

An abstract FlexcomFunction class is created that will be extended for
specific functions. Each function has a dedicated overlapped
MemoryRegion that it is going to be activated when that particular
function is selected.

A generic MemoryRegion is used to map generic registers (PSELID, PID)
that is always active but has a lower priority. The function memory
region is trimed to not include the generic registers.

The FlexcommFunction object also gets access to the (non-concurrently)
shared register backstore.

The patch includes an automatically generated header which contains
the register layout and helpers.

The header can be regenerated with the svd-flexcomm target when the
build is configured with --enable-mcux-soc-svd.

GitWatcher: ignore
Change-Id: I2429bc497c637cc7a3a7e2b60d896b263e47d494
Signed-off-by: Octavian Purdila <tavip@google.com>