hw/arm: add basic support for the RT500 SoC
Add basic support for the RT500 SoC. It supports enough peripherals to
run the NXP's microXpresso SDK hello world example.

The patch includes an automatically generated header which contains
peripheral base addreses and interrupt numbers.

The header can be regenerated with the svd-rt500 target when the
build is configured with --enable-mcux-soc-svd.

GitWatcher: ignore
Change-Id: I9b1aba3cd9a30c048d61c16b96038e25229fcb64
Signed-off-by: Octavian Purdila <tavip@google.com>