Implement OOB Access to goldfish address space.
We observed that some times the guest could read/write to an address
that is not backed by RAM at the moment of access. It is likely a
wrong address from guest graphics stack when Vulkan is enabled.
However, it is hard to find out where.

Let's ignore these OOB accesses to goldfish address space.

Change-Id: I9c472a78d9b9e12167a039814adaf90c570dafa1