Add function to get CPU brand name (and core count) using WMI on Windows.
This is Windows only. But it should work on x86 and ARM64 architectures.
The existing core count implementation is actually logical processor
count. It could be decided later whether we will switch to using real
core count. However, getting core count uses the same approach as
getting brand name. Thus, it does not harm anything to get this
additional information for future use.

Change-Id: Ib2c9c07ecd12c38583224ff0c929ea2b7949c0f7