Simplify XrOptions event notification
* Send XrOptions event received from xrOptionsPublisher instead of calling getXrOptionsNotificationEvent

*** Why this change? ***
Inside `registerOnce`, the function ptr already received event as parameter which can be directly passed to mNotificationListeners to fireEvent.
I was confused also why in the orignal code this argument is being ignored and calling getEvent (e.g, getBootedNotificationEvent or getXrOptions again).
This pattern to ignore received notification event and query event again
is unnecessary and less performant. So I made the refactor.
I would like to update the registerOnce for other publishers to follow the same pattern.

*** Why the first event in `notificationStream` needs to call `getXrOptions`. ***
Regarding why the first event in `notificationStream` needs to call `getXrOptions`.
This is because `notificationStream` does not have event as parameter.
```
NotificationStreamWriter* NotificationStream::notificationStream()
```

Bug: b/400473227
Test: click eye icon on Studio and guest toggle button, verified Studio received notification in log. demo video: https://drive.google.com/file/d/1QG0gXc5kliSvBousBaWoclXs04bmRKxP/view?usp=drive_link

Test environment:
* Mac M1 Pro
* Android Studio Meerkat Feature Drop | 2024.3.2 Nightly 2025-03-12
* Guest system image: ab/P92927767 (CL pending review:http://nfnt/q/topic:%22studio-notification%22)
* launch command: `objs/emulator -avd XR_Device -dns-server 8.8.8.8 -feature GuestAngle -qt-hide-window -grpc-use-token -idle-grpc-timeout 300 -verbose`

Change-Id: I7b9d154ec4662ac0026c56536e99cfd43e9631e0