Enable virtual queue by default
The feature flag needs to be initialized before the renderer
initialization to take affect. But the boot parameter setup
is done after the feature flags are consumed. Setting up the
flag at runtime based on capabilities and AVD settings
requires some overhaul on the renderer config and startup,
so we'll enable it by default instead to avoid infinite boot
loading issues on XR AVDs.

Bug: 407751250
Test: run XR emulator
Change-Id: I6c9969aa0a17726e9532620ed22bbe106bf9bf48