hw/i3c/mipi-hci: Add DMA register definitions
The DMA (or ring) consists of header registers, and status registers.

The ring header describes the size of the structures used in DMA
transfers, and also advertises how many different rings there are.
Currently we only model 1 ring.

The status registers are used to control the DMA and handle
transactions.

GitWatcher: ignore
Google-Bug-Id: 406093016
Change-Id: I994d0d4845caf491b8413818d49c562f5d303e58
Signed-off-by: Joe Komlodi <komlodi@google.com>