hw/i3c/mipi-hci: dma: Add Address command support
This command does either ENTDAA or SETDASA.

The way that this works is the driver will populate the the DAT with the
addresses that it wants to assign to targets on the bus. Once that's
done, it creates the DMA descriptor that says to do an address
assignment command.

The address assignment command itself, like other commands, is the same as
it would be as if it were initiated by PIO, so it's factored into the
command file.

Once the command is done, the response indicating how the command went
(how many devices were assigned, did it succeed or get NACKed, etc) is
pushed onto the DMA response ring buffer.

GitWatcher: ignore
Google-Bug-Id: 406094038
Change-Id: If6d63fca79e26186ba64bd53c8fea8c8f4ab5a09
Signed-off-by: Joe Komlodi <komlodi@google.com>