hw/i3c/mipi-hci: dma: Add IBI handling
Largely similar to how other transactions are stored. Once an IBI is
handled, we write the IBI status (who IBI'd, how was it handled, etc),
and the IBI data (such as MDBs) to separate IBI status and data ring
buffers respectively.

Once those are stored, we update the IBI status and interrupt registers
to reflect that state.

GitWatcher: ignore
Google-Bug-Id: 406094058
Change-Id: Ice0c3cda458278344623eb6638ded49494b6e07d
Signed-off-by: Joe Komlodi <komlodi@google.com>