hw/i3c/mipi-hci: Add halt support
On a transaction error, the controller enters a halt state. The MIPI HCI
spec says this can be recorded in an extended capabilities status register,
but it's not required. Because of this, we make the enter_halt function
overridable.

When halted, the controller cannot do any transfers until the user writes
to HC_CONTROL.RESUME

GitWatcher: ignore
Google-Bug-Id: 406092085
Change-Id: I74985e225b366d4c21b3d3565cc83fa1b860ae7f
Signed-off-by: Joe Komlodi <komlodi@google.com>