hw/i2c/npcm_smbus: Send a STOP when disabled
When the NPCM SMBus module gets disabled, it stops SCL and transmitting
data, bringing both lines high. This can be interpreted as an I2C STOP
from the target.

In practice, what happens is when Linux thinks the bus is stuck, it will
disable and re-enable the bus to get it into a known state. If this
happens in the middle of a transaction with an SMBus device, the SMBus
device is left in its old state. This means when a new transaction
comes, the SMBus device is now confused and will become unresponsive
until reset.

To fix this, we'll model the I2C stop that's incidentally sent when the
controller is disabled.

GitWatcher: ignore
Google-Bug-Id: 231153633
Signed-off-by: Joe Komlodi <komlodi@google.com>
Change-Id: I9bdf8d468243a8d1ac38f898fe1d46e34260b9be