hw/i3c/ast27xx-i3c: Add get next dynamic address handler
This overrides the MIPI HCI implementation, since the way it works on
the Aspeed controller is different.

The driver will set the corresponding bit in I3C_DAA_INDEX_n for the
address it wants to set. For example, to set address 9, the 9th bit in
I3C_DAA_INDEX_0 is set. For address 32, the 0th bit in I3C_DAA_INDEX_1
is set.

It's unclear what happens if multiple bits are set, so we assume that
the lowest bit set is the one used for the address.

GitWatcher: ignore
Google-Bug-Id: 406093286
Change-Id: Ic2db4a5308d8dfe4e16a8bb2a7a3ca58c993cf58
Signed-off-by: Joe Komlodi <komlodi@google.com>