hw/i3c/ast27xx-i3c: Add get device dynamic address function
This overrides the MIPI HCI one for the Aspeed-specific implementation.

In the HCI implementation, the dynamic address is retrieved from
TARGET_DAT.TARGET_DYNAMIC_ADDRESS in the DAT table, where the index into
the DAT table is determined from the transfer descriptor.
In the Aspeed implementation, the dynamic address is the index from the
transfer descriptor.

The index passed in is the index into the table, of which every DAT
entry is 2 32-bit registers, so divide the DAT index by 2 to get the
dynamic address.

GitWatcher: ignore
Google-Bug-Id: 406093286
Change-Id: I071e517099a70e7e3a2a93dc2774a482f4e63aa9
Signed-off-by: Joe Komlodi <komlodi@google.com>