Disable virtual queue for XR AVDs on Mac
Virtual queue and subsequent enablement of skiavk is causing
kernel lockups on goldfish_sync and blocking user logins.
Running with skiagl with virtual queue enabled, or only enabling
virtual queue only for HWUI doesn't fix the issue.
This change disables the virtual queue usage on Mac for XR AVDs
to de-escalate the issue until a proper fix is made.

Bug: 404619245
Test: run XR emulator, add account
Change-Id: I38a7ebcf7f4ab7d4865ffb1b94fde2bf2ececfb8