Revert "Revert "Revert "WHPX: cpuid will return qemu settings instead of...""
Reason of Original CL:
The original CL was made to let WHPX accept CPU profile definitions as
KVM did. And the reason of doing so was related to the effort to make
a snapshot less dependent to the OSes or hypervisors so that a snapshot
saved on one OS or hypervisor can be loaded on another. What's more,
the default CPU profile should contain as few features as possible
(only those features that all hypervisors we work on can support). This
resulted in AES and F16C being removed by another CL (submitted together
with this CL).

Reason of Revert:
1. ndk translation requires AES and F16C. We need to include these two
   in the default CPU profile. This breaks the purpose of the original
   CLs.
2. The original CL achieved its purpose by intercepting the guest CPUID
   instruction. It intercepted way more CPUID leaves than the initial
   WHPX implementation and the current upstream QEMU WHPX implementation
   which is made and maintained by Microsoft. Creating a departure from
   Microsoft's approach brings maintenance overhead. And Microsoft may
   not be able to help us in case of errors.
3. We did not work on the snapshot improvement mentioned above any more.
So it is better to revert the CL so that WHPX works as intended by the
Microsoft.

Bug: 388718417

This reverts commit e37bda0b5bfa6f9f2d242e75d3176a0921242cd5.

Change-Id: I111b8e2f1737e7cb5a9016a6db1d71efddcd68ae