fix(build): Add prebuilt ninja to PATH
This change ensures that the version of Ninja used for building the
emulator is the one prebuilt for the current platform. By explicitly
adding the path to the prebuilt Ninja (located at
`prebuilts/ninja/{platform}-x86`) to the environment PATH, we remove
a dependency on potentially different versions of Ninja, such as the one
bundled with Visual Studio's CMake package on Windows machines.

This leads to a more consistent and predictable build environment across
different developer setups.

Change-Id: I121537aa0d8f488b553b97f12fef0b7895d4f702