build: Enforce macOS SDK 13.3 for toolchain
This change updates the build configuration to strictly enforce the use
of macOS SDK 13.3. This is necessary due to compatibility issues
observed between Clang and newer SDK versions.

Enforcing a specific SDK version aims to create a "pseudo-hermetic"
toolchain, ensuring that developer environments align with the buildbot
setup. This consistency helps prevent and diagnose compatibility
problems more effectively.

Developers will now be required to install macOS SDK 13.3 if it is not
already present. A new script `get-osx-sdk.sh` has been added to
facilitate this process.

Bug: 419844119
Change-Id: I2516be767910e34aa75aa7f31dcd3666a7db9c4f