hw/sd: Fix npcm7xx_sdhci test invalid reads
A future change will cause qtest_readl() to abort when accessing an
invalid register. That change would cause the 'npcm7xx_sdhci/reset'
test to fail because it accesses registers which were invalid.

This change does the following:

1. Relaxes npcm7xx_sdhci_check_mem_op() to allow reads of undefined
   registers in the overlay it defines at offset NPCM7XX_PRSTVALS (0x60)
   with size NPCM7XX_SDHCI_REGSIZE (0x20).

2. Updates the 'npcm7xx_sdhci/reset' test to specifically read the
   BOOTTOCTR register as a 32-bit value.

   Note that NPCM7XX_BOOTTOCTR (0x10) is an offset from the start of the
   overlay at NPCM7XX_PRSTVALS (0x60), which should probably be named
   something like NPCM7XX_CUSTOM_REGS.

3. Replaces NPCM7XX_REG_SIZE with SDHC_REGISTERS_MAP_SIZE for clarity.

This was tested by cherry-picking the change to qtest_process_command()
which adds error-checking to the address_space_read() and
address_space_write() calls.

GitWatcher: ignore
Google-Bug-Id: 421314248
Change-Id: I2dd5bee2d877316228e52c78fe26fc70d90b795c
Signed-off-by: Jonathon Reinhart <jrreinhart@google.com>