Applying a similar fix as https://android-review.git.corp.google.com/c/platform/hardware/generic/goldfish/+/3197386 Run command unit tests as deamon
It is possible in linux for the process history to be cleaned out before
we got a reference to process. We now run the test process as a deamon
to make sure the process is always alive when we are querying it.
We have seen this test failing on some machine with a 259 (STILL_ALIVE) status

Bug: 421266975
Change-Id: Ic9ac144cf913ca6f8a3eaa64d4297a1a8c42352e