Avoid libunwind warnings from lavapipe
This is because the lavapipe driver expects libunwind support from the
underlying c library. Our libc++ lib is incompatible with the lavapipe
request resulting in log spams

We can get around it by preloading libgcc if it's present on the system.
We will need to revisit this to provide a better solution but for now,
the situation is improved.

If the library is not present on a system, the only issue we'll see is
the error logs coming back.

Bug: 417138854
Change-Id: I2b624b988ae1375c1a298406bdbe01af1ef239eb