hw/pci-host/npcm_pcierc: Unmap before resetting
If we restart the machine, the AXI2PCIe and PCIe2AXI windows are zeroed
out, but are still mapped onto system memory. This causes problems when
the BMC tries to re-map the windows, since it looks like the windows are
unmapped, but they actually aren't.

To fix this, we'll just explicitly zero out each field, unmap the
window, and then zero out the window memory region.

GitWatcher: ignore
Google-Bug-Id: 427748383
Change-Id: I49caeeaa7ad490773081b31106d956e4c1a85ddc
Signed-off-by: Joe Komlodi <komlodi@google.com>