Build MoltenVK with emulator build.
You can run just the prebuilts task manually:
> python3 android/build/python/cmake.py --task Prebuilts

In a full emulator build, the prebuilt SHA1s will be compared
against the git repository. If there's a difference, then that
prebuilt will be rebuilt and the binaries will be dropped into
it's corresponding prebuilts directory. Then the emulator build
will use the resulting binaries.

If for some reason, we need to retrigger a build and the SHA1 has
not changed, we can simply create a CL in the prebuilts to delete
the SHA1 file.

Bug: 427992749
Test: presubmit
Change-Id: I45e1c726353548092c716695c3954bff56cdc28f