hw/i2c/aspeed_i2c: Add DMA xfer byte hardware bug
Currently, the driver looks at the buffer status registers to see what
byte was last received, if the transfer is one where the first byte
received determines the length of the rest of the transfer.
The datasheet explicitly says that this register is not used in DMA
transactions, and there are other ways to read data from a DMA
transaction, such as reading the memory. However, the driver uses this
register anyway to read the length byte from a transaction.

Since this driver works on hardware, it likely means that this register
does actually have behavior in DMA mode, despite the register naming and
documentation suggesting otherwise, so model the incorrect behavior for
now so the driver can work.

GitWatcher: ignore
Google-Bug-Id: 418905237
Change-Id: I991ee64b8bca040a6a54775dbb553878613f618f
Signed-off-by: Joe Komlodi <komlodi@google.com>