feat: Add error messages to crash reports
This change ensures that `LOG(ERROR)` messages are included in crash
reports. A new `CrashpadLogSink` is introduced, which registers itself
with the logging framework to capture and append error messages to the
crash report. This uses a circular buffer of 8KB to store the messages,
overwriting older entries as needed.

The change includes a unit test to verify that error messages are
correctly captured and present in the generated minidump.

Sample snippet from a report:

 "annotation_objects": [
      {
        "name": "ERRLOG",
        "value": "Unable to connect to adb daemon on port: 5037\n"
      },

Change-Id: I2388026115f21775c28ccaeb62314df0c390fe86