Add microphone toggle button to the toolbar UI
This change adds the icon to the toolbar UI and hooks up the code callbacks for the button click.
In this change, clicking the button doesn't yet do anything; the functionality will be added in subsequent changes.
While this feature is under development, it is gated behind the `MicrophoneToggleUI` feature flag.

Eventually, this toggle will enable/disable the microphone input for the emulator. This is the same as the
button "Virtual microphone uses host audio input" in the extended controls panel. Further changes will
update the extended controls panel based on this change.

Test: Tested with:
1. `-feature MicrophoneToggleUI` -> Microphone icon appears in toolbar
2. `-feature -MicrophoneToggleUI` -> Microphone icon is hidden
3. Feature flag omitted -> Microphone icon is hidden

Bug: b/440383173
Change-Id: I6f4a5a047f3d12c537ea2428a0e7bcf692d06739