Add touchpad pane to extended panel
Adds the touchscreen widget to an extended panel page for now.
The touchpad is scaled to fit on the extended window properly, and will
send its events to the touchpad agent.

The page is sized dynamically, so should be suitable for displaying
elsewhere if we decide to move the touchpad's location.

Bug: 441089953
Change-Id: Ia6c0c996a218f7b411096350bd2df74739e2a360