Hook up microphone toggle to hardware input
Connect the microphone toggle button to enable/disable routing the host
microphone input into the AVD. This mirrors how the microphone-page.cpp
extended page works. Note that this does NOT mirror the values between
the two pages; this will be addressed in a future change.

While this feature is under development, it is gated behind the
`-feature MicrophoneToggleUI` feature flag.

Bug: b/440383173
Fixes: b/445494473
Test: Check toggle value is persisted through reboot.
Test: Test microphone input is toggled on and off via the google voice assistant
Change-Id: Iba88a28d7d0a405b289cc037a143f818b7012127