Dock the touchpad for standalone glasses emulator
Displays the touchpad docked below the main screen for glasses when the
emulator is run in standalone mode. For embedded mode, the touchpad is
still displayed in the extended controls.

Key events are forwarded to the tool window for processing so hotkeys
will still work while using the touchpad.

This also adds keyrelease forwarding for the extended controls that was
previously missing.

Bug: 446245406
Change-Id: If55042eff9773d3cf45f0d3a53be5538ecdd25f0