Fix HVF crashes on MacOS 12 or lower.
Starting from MacOS 13, size of guest physical address space can be
more than 2^36. New hvf APIs are introduced to set the desired size.

For MacOS 12 and lower, we have to stick with the limit of 36 bits.
Thus, we also need to change generic QEMU code as the assumption is
at least 40 bits when it partitions guest physical address space for
RAM, device memory, etc.

As a result of this, it is desired to revert this CL when MacOS 12
or lower is not supported any more.

Bug: 453646948

Change-Id: I82357b0de61bfc47f8ce591c4283d65167608714