fix(macos): Allow compilation with MacOSX SDK 26.0
Add MacOSX SDK version 26.0 to the list of supported SDKs in the
`gen-android-sdk-toolchain.sh` script. This enables building the
project with recent versions of Xcode.

Additionally, suppress linker warnings (`-Wl,-w`) during the macOS
build process. This is necessary because the build currently uses
older prebuilt libraries, which generate warnings with newer SDKs and
linkers.

Change-Id: I0b21f5ca9766e1aacc84396925d800acaeac4d57